+++ common/src/main/java/net/satisfy/wildernature/WilderNature.java
@@ -1,8 +1,11 @@
 package net.satisfy.wildernature;
 
+import dev.architectury.registry.ReloadListenerRegistry;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.PackType;
 import net.satisfy.wildernature.core.event.BountyEvents;
 import net.satisfy.wildernature.core.event.VanillaBoneThrowEvent;
+import net.satisfy.wildernature.core.fieldguide.FieldGuideDataLoader;
 import net.satisfy.wildernature.core.network.BountyBoardNetworking;
 import net.satisfy.wildernature.core.registry.*;
 
@@ -26,6 +29,7 @@ public class WilderNature {
         RecipeRegistry.init();
         TabRegistry.init();
         SoundEventRegistry.init();
+        ReloadListenerRegistry.register(PackType.SERVER_DATA, FieldGuideDataLoader.INSTANCE);
     }
 }